// generated/proto/apigateway_pb.ts
export class ApiGatewayRequest extends Message<ApiGatewayRequest> {
  /**
   * Имя сервиса для вызова
   *
   * @generated from field: string service = 1;
   */
  service = "";

  /**
   * Имя метода для вызова
   *
   * @generated from field: string method = 2;
   */
  method = "";

  /**
   * Тело запроса в виде динамической структуры
   *
   * @generated from field: google.protobuf.Struct request_body = 3;
   */
  requestBody?: Struct;

  constructor(data?: PartialMessage<ApiGatewayRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "apigateway.ApiGatewayRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "service", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "method", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "request_body", kind: "message", T: Struct },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): ApiGatewayRequest {
    return new ApiGatewayRequest().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): ApiGatewayRequest {
    return new ApiGatewayRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): ApiGatewayRequest {
    return new ApiGatewayRequest().fromJsonString(jsonString, options);
  }

  static equals(
    a: ApiGatewayRequest | PlainMessage<ApiGatewayRequest> | undefined,
    b: ApiGatewayRequest | PlainMessage<ApiGatewayRequest> | undefined,
  ): boolean {
    return proto3.util.equals(ApiGatewayRequest, a, b);
  }
}

/**
 * Ответ от API Gateway
 *
 * @generated from message apigateway.ApiGatewayResponse
 */
export class ApiGatewayResponse extends Message<ApiGatewayResponse> {
  /**
   * Результат в виде either/or: либо тело ответа, либо ошибка
   *
   * @generated from oneof apigateway.ApiGatewayResponse.result
   */
  result:
    | {
        /**
         * Тело ответа в виде динамической структуры
         *
         * @generated from field: google.protobuf.Struct response_body = 1;
         */
        value: Struct;
        case: "responseBody";
      }
    | {
        /**
         * Информация об ошибке
         *
         * @generated from field: apigateway.Error error = 2;
         */
        value: Error;
        case: "error";
      }
    | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<ApiGatewayResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "apigateway.ApiGatewayResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    {
      no: 1,
      name: "response_body",
      kind: "message",
      T: Struct,
      oneof: "result",
    },
    { no: 2, name: "error", kind: "message", T: Error, oneof: "result" },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): ApiGatewayResponse {
    return new ApiGatewayResponse().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): ApiGatewayResponse {
    return new ApiGatewayResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): ApiGatewayResponse {
    return new ApiGatewayResponse().fromJsonString(jsonString, options);
  }

  static equals(
    a: ApiGatewayResponse | PlainMessage<ApiGatewayResponse> | undefined,
    b: ApiGatewayResponse | PlainMessage<ApiGatewayResponse> | undefined,
  ): boolean {
    return proto3.util.equals(ApiGatewayResponse, a, b);
  }
}

/**
 * Информация об ошибке
 *
 * @generated from message apigateway.Error
 */
export class Error extends Message<Error> {
  /**
   * Короткое сообщение об ошибке
   *
   * @generated from field: string short_message = 1;
   */
  shortMessage = "";

  /**
   * Подробное описание ошибки
   *
   * @generated from field: string description = 2;
   */
  description = "";

  /**
   * Код ошибки
   *
   * @generated from field: int32 code = 3;
   */
  code = 0;

  /**
   * Дополнительные детали ошибки
   *
   * @generated from field: repeated string details = 4;
   */
  details: string[] = [];

  constructor(data?: PartialMessage<Error>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "apigateway.Error";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    {
      no: 1,
      name: "short_message",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
    },
    {
      no: 2,
      name: "description",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
    },
    { no: 3, name: "code", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
    {
      no: 4,
      name: "details",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
      repeated: true,
    },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): Error {
    return new Error().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): Error {
    return new Error().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): Error {
    return new Error().fromJsonString(jsonString, options);
  }

  static equals(
    a: Error | PlainMessage<Error> | undefined,
    b: Error | PlainMessage<Error> | undefined,
  ): boolean {
    return proto3.util.equals(Error, a, b);
  }
}